chore: move fixtures out of conftest
